Renamed two views to make the naming convetions streamlined.
Also decided to only show user decks in the deck history view.
