+++ Grand Archive Companion/Grand Archive Companion/Views/Deck/DeckHistoryView.swift
@@ -1,78 +1,81 @@
 //
-//  DeckHistoryView.swift
+//  DeckListView.swift
 //  Grand Archive Companion
 //
-//  Created by Tarik Fayad on 1/28/25.
+//  Created by Tarik Fayad on 2/3/25.
 //
 
 import SwiftUI
 
 struct DeckHistoryView: View {
     @Environment(\.presentationMode) var presentationMode
+    @Environment(\.colorScheme) var colorScheme
     @Environment(\.modelContext) private var modelContext
-    
-    @State var matches: [Match] = []
-    @State var filteredMatches: [Match] = []
-    @State var decks: [Deck] = []
-    
-    @State private var menuTitle: String = "Filter by Deck"
+
+    @State private var debouncedSearch: (() -> Void)?
+    @State private var searchText: String = ""
+    @State private var decks: [Deck] = []
+    @State private var filteredDecks: [Deck] = []
+    @State private var isLoading: Bool = false
+    @State private var navigateToDeckView = false
     @State private var selectedDeck: Deck?
-    @State private var navigateToCreateMatchView = false
-    
+
     var body: some View {
-        GeometryReader { geometry in
-            ZStack {
-                Color.background.ignoresSafeArea(.all)
-                VStack {
-                    Spacer()
-                    Menu(menuTitle) {
-                        Button("Clear Filter") {
-                            menuTitle = "Filter by Deck"
-                            filteredMatches = matches
-                        }
-                        ForEach(decks, id: \.self) { deck in
-                            Button(deck.name) {
-                                filteredMatches = filterMatchesByDeck(deck)
-                                menuTitle = deck.name + " Results"
-                                selectedDeck = deck
-                            }
-                        }
-                    }.frame(width: max(geometry.size.width - 30, 0), height: 35)
-                        .background(Color.secondary)
-                    
-                    HStack {
-                        Text("You")
-                            .frame(width: max(geometry.size.width / 2 - 20, 0), height: 35)
-                            .font(.system(size: 20, weight: .bold, design: .default))
-                            .background(.playerBlue)
-                        
-                        Text("Opponent")
-                            .frame(width: max(geometry.size.width / 2 - 20, 0), height: 35)
-                            .font(.system(size: 20, weight: .bold, design: .default))
-                            .background(.playerPink)
-                    }
-                    
-                    if menuTitle != "Filter by Deck" {
-                        let percent = String(format: "%.2f", Deck.winRate(deck: selectedDeck!, context: modelContext))
-                        Text("\(percent)% Win Rate")
-                            .font(.caption)
-                            .textCase(.uppercase)
-                    }
-                    
-                    List {
-                        ForEach(filteredMatches, id: \.self) { match in
-                            MatchRowView(match: match)
-                        }
+        ZStack {
+            Color.background.ignoresSafeArea(.all)
+            
+            List(filteredDecks, id: \.self) { deck in
+                if isLoading {
+                    ProgressView("Searching for decks...")
+                } else {
+                    DeckRowView(deck: deck)
                         .listRowBackground(Color.background)
-                    }
-                    .scrollContentBackground(.hidden)
-                    .background(Color.background)
-                    .padding(.horizontal, -20)
+                        .onTapGesture {
+                            selectedDeck = deck
+                            navigateToDeckView = true
+                        }
                 }
-            }.foregroundStyle(.white)
+            }
+            .scrollContentBackground(.hidden)
+            .searchable(text: $searchText, placement: .navigationBarDrawer, prompt: "Filter by deck name or champion...")
+            .disableAutocorrection(true)
+            .onSubmit(of: .search) {
+                 filterDecks()
+            }
+            .onChange(of: searchText) { newValue in
+                // MARK: Deprecated but much easier to use than alternatives. Will change when needed.
+                searchText = newValue
+                debouncedSearch?()
+            }
+            .overlay {
+                if decks.isEmpty {
+                    ContentUnavailableView("Decks", systemImage: "magnifyingglass.circle", description: Text("You haven't saved any decks yet."))
+                        .foregroundStyle(.white)
+                }
+            }
         }
-        .navigationDestination(isPresented: $navigateToCreateMatchView){
-            AddMatchView()
+        .onAppear {
+            let appearance = UINavigationBarAppearance()
+            appearance.configureWithOpaqueBackground()
+            appearance.backgroundColor = .background
+            UINavigationBar.appearance().standardAppearance = appearance
+            UINavigationBar.appearance().scrollEdgeAppearance = appearance
+            
+            if colorScheme != .dark {
+                UITextField.appearance(whenContainedInInstancesOf: [UISearchBar.self]).textColor = .white
+            }
+            
+            debouncedSearch = debounce(delay: 0.5) {
+                 filterDecks()
+            }
+            
+            decks = Deck.loadUserDecks(context: modelContext)
+            filteredDecks = decks
+        }
+        .navigationDestination(isPresented: $navigateToDeckView) {
+            if let selectedDeck = selectedDeck {
+                DeckDetailView(deck: selectedDeck)
+            }
         }
         .navigationBarBackButtonHidden(true)
         .toolbar {
@@ -82,63 +85,42 @@ struct DeckHistoryView: View {
                 }) {
                     HStack {
                         Image(systemName: "arrow.backward") // Custom back icon
-                    } .foregroundStyle(.white)
-                }
-            }
-            
-            ToolbarItem(placement: .principal) {
-                Text("Match History")
+                    }
                     .foregroundStyle(.white)
-                    .fontWeight(.bold)
-            }
-            
-            ToolbarItem(placement: .navigationBarTrailing) {
-                Button(action: {
-                    // Allow user to save a match
-                    navigateToCreateMatchView = true
-                }) {
-                    HStack {
-                        Image(systemName: "plus") // Custom back icon
-                    } .foregroundStyle(.white)
                 }
             }
-        }
-        .navigationBarTitleDisplayMode(.inline)
-        .onAppear {
-            let appearance = UINavigationBarAppearance()
-            appearance.configureWithTransparentBackground()
-            UINavigationBar.appearance().standardAppearance = appearance
-            UINavigationBar.appearance().scrollEdgeAppearance = appearance
-            
-            matches = Match.load(context: modelContext)
-            filteredMatches = matches
-            
-            var myDecks: Set<Deck> = []
-            for match in matches {
-                myDecks.insert(match.userDeck)
-            }
-            decks = Array(myDecks).sorted { $0.name < $1.name }
         }
     }
     
-    private func filterMatchesByDeck(_ deck: Deck?) -> [Match] {
-        guard let deck else {
-            return matches
+    func filterDecks() {
+        Task {
+            isLoading = true
+            filteredDecks = decks.filter { deck in
+                deck.name.contains(searchText) || deck.champions.contains(where: { $0.name.contains(searchText) })
+            }
+            if searchText == "" { filteredDecks = decks }
+            isLoading = false
         }
-        return matches.filter { $0.userDeck == deck }
+    }
+    
+    func debounce(delay: TimeInterval, action: @escaping () -> Void) -> () -> Void {
+        var lastFireTime = DispatchTime.now()
+        let queue = DispatchQueue.main
+
+        return {
+            let now = DispatchTime.now()
+            let when = DispatchTime.now() + delay
+            lastFireTime = now
+
+            queue.asyncAfter(deadline: when) {
+                if lastFireTime == now {
+                    action()
+                }
+            }
+        }
     }
 }
 
 #Preview {
-    let playerChampion = Champion.init(name: "Lorraine, Crux Knight", lineage: "Lorraine", jobs: ["Warrior"], health: 28, level: 3)
-    let playerDeck = Deck(name: "Deck 1", isUserDeck: true, champions: [playerChampion], elements: [.fire, .crux])
-    
-    let opponentChampion = Champion.init(name: "Lorraine, Crux Knight", lineage: "Lorraine", jobs: ["Warrior"], health: 28, level: 3)
-    let opponentDeck = Deck(name: "Deck 1", isUserDeck: false, champions: [opponentChampion, playerChampion], elements: [.fire, .crux, .water])
-    
-    let match = Match.init(didUserWin: true, userDeck: playerDeck, opponentDeck: opponentDeck)
-    
-    NavigationStack {
-        DeckHistoryView(matches: [match, match, match, match, match, match, match, match, match, match])
-    }
+    DeckHistoryView()
 }
